+++ CHANGELOG.md
@@ -7,6 +7,48 @@
 * Unity: add methods for setting autoNotify and autoDetectAnrs
   [#1233](https://github.com/bugsnag/bugsnag-android/pull/1233)
 
+* Including bugsnag.h in C++ code will no longer cause writable-strings warnings
+  [1260](https://github.com/bugsnag/bugsnag-android/pull/1260)
+
+## 5.9.3 (2021-05-18)
+
+* Avoid unnecessary collection of Thread stacktraces
+  [1249](https://github.com/bugsnag/bugsnag-android/pull/1249)
+
+* Prevent errors in rare cases where either ConnectivityManager or StorageManager is not available
+  [1251](https://github.com/bugsnag/bugsnag-android/pull/1251)
+  
+* Change the Bugsnag-Internal-Error header to "bugsnag-android"
+  [1252](https://github.com/bugsnag/bugsnag-android/pull/1252)
+
+* Prevent resource exhaustion when Throwable cause chains are recursive
+  [1255](https://github.com/bugsnag/bugsnag-android/pull/1255)
+
+* Added Date support to ObjectJsonStreamer
+  [1256](https://github.com/bugsnag/bugsnag-android/pull/1256)
+
+## 5.9.2 (2021-05-12)
+
+### Bug fixes
+
+* Guard against exceptions with null stack traces
+  [#1239](https://github.com/bugsnag/bugsnag-android/pull/1239)
+
+* Fix bug that terminated the app when multiple ANRs occur
+  [#1235](https://github.com/bugsnag/bugsnag-android/pull/1235)
+
+* Prevent rare NPE in log message
+  [#1238](https://github.com/bugsnag/bugsnag-android/pull/1238)
+
+* Prevent rare NPE when capturing thread traces
+  [#1237](https://github.com/bugsnag/bugsnag-android/pull/1237)
+
+* Catch exceptions thrown by Context.registerReceiver to prevent rare crashes
+  [#1240](https://github.com/bugsnag/bugsnag-android/pull/1240)
+
+* Fix possible NegativeArraySizeException in crash report deserialization
+  [#1245](https://github.com/bugsnag/bugsnag-android/pull/1245)
+
 ## 5.9.1 (2021-04-22)
 
 ### Bug fixes